moveit::move_ref

Struct MoveRef

pub struct MoveRef<'a, T: ?Sized> { /* private fields */ }

A MoveRef<'a, T> represents an owned T whose storage location is valid but unspecified.

See the module documentation for more details.

Implementations

impl<'a, T: ?Sized> MoveRef<'a, T>

pub unsafe fn new_unchecked(ptr: &'a mut T, drop_flag: DropFlag<'a>) -> Self

Creates a new MoveRef<T> out of a mutable reference.

Safety

ptr must satisfy the longest-lived criterion: after the return value goes out of scope, ptr must also be out-of-scope. Calling this function correctly is non-trivial, and should be left to moveit!() instead.

In particular, if ptr outlives the returned MoveRef, it will point to dropped memory, which is UB.

drop_flagโs value must not be dead, and must be a drop flag governing the destruction of ptrโs storage in an appropriate manner as described in moveit::drop_flag.

pub fn into_pin(this: Self) -> Pin<Self>

Converts a MoveRef<T> into a Pin<MoveRef<T>>.

Because we own the referent, we are entitled to pin it permanently. See Box::into_pin() for a standard-library equivalent.

pub fn as_ptr(this: &Self) -> *const T

Returns this MoveRef<T> as a raw pointer, without creating an intermediate reference.

The usual caveats for casting a reference to a pointer apply.

pub fn as_mut_ptr(this: &mut Self) -> *mut T

Returns this MoveRef<T> as a raw mutable pointer, without creating an intermediate reference.

The usual caveats for casting a reference to a pointer apply.

impl<'a, T> MoveRef<'a, T>

pub fn into_inner(this: Self) -> T

Consume the MoveRef<T>, returning the wrapped value.