moveit::new
Trait Swap

pub trait Swap<Rhs = Self> {
    // Required method
    fn swap_with(self: Pin<&mut Self>, src: Pin<&mut Rhs>);
}

A swappable type, which is able to efficiently swap the contents of two of its values.

Unlike New, Swap is safe, because it does not impose any requirements on the swapped pointers.

It is possible to implement swapping with a source type that isn’t Self.

Required Methods

fn swap_with(self: Pin<&mut Self>, src: Pin<&mut Rhs>)

Swaps the contents of self and src without running any destructors.