moveit::slot

Struct DroppingSlot

pub struct DroppingSlot<'frame, T> { /* private fields */ }

Similar to a Slot, but able to drop its contents.

A DroppingSlot wraps a Slot, and will drop its contents if the Slot’s drop flag is dead at the time of the DroppingSlot’s destruction.

This type has an API similar to Slot’s, but rather than returning MoveRefs, which would own the contents of this slot, we return a &mut T and a DropFlag, which the caller can assemble into an appropriately-shaped MoveRef. The drop flag will be one decrement away from being dead; callers should make sure to decremement it to trigger destruction.

DroppingSlot is intended to be used with DerefMove::deref_move(), and will usually not be created by moveit’s users. However, slot!() provides DroppingSlot support, too. These slots will silently forget their contents if the drop flag is left untouched, rather than crash.

Implementations

impl<'frame, T> DroppingSlot<'frame, T>

pub unsafe fn new_unchecked( ptr: &'frame mut MaybeUninit<T>, drop_flag: DropFlag<'frame>, ) -> Self

Creates a new DroppingSlot with the given pointer as its basis.

To safely construct a DroppingSlot, use slot!().

Safety

ptr must not be outlived by any other pointers to its allocation.

drop_flag’s value must be dead, and must be a drop flag governing the destruction of ptr’s storage in an appropriate manner as described in moveit::drop_flag.

pub fn put(self, val: T) -> (&'frame mut T, DropFlag<'frame>)

Put val into this slot, returning a reference to it.

pub unsafe fn pin(self, val: T) -> (Pin<&'frame mut T>, DropFlag<'frame>)

Pin val into this slot, returning a reference to it.

Safety

This function pins the memory this slot wraps, but does not guarantee its destructor is run; that is the caller’s responsibility, by decrementing the given DropFlag.

pub unsafe fn emplace<N: New<Output = T>>( self, new: N, ) -> (Pin<&'frame mut T>, DropFlag<'frame>)

Emplace new into this slot, returning a reference to it.

Safety

This function pins the memory this slot wraps, but does not guarantee its destructor is run; that is the caller’s responsibility, by decrementing the given DropFlag.

pub unsafe fn try_emplace<N: TryNew<Output = T>>( self, new: N, ) -> Result<(Pin<&'frame mut T>, DropFlag<'frame>), N::Error>

Try to emplace new into this slot, returning a reference to it.

Safety

This function pins the memory this slot wraps, but does not guarantee its destructor is run; that is the caller’s responsibility, by decrementing the given DropFlag.