moveit::slot

Struct Slot

pub struct Slot<'frame, T> { /* private fields */ }

An empty slot on the stack into which a value could be emplaced.

The 'frame lifetime refers to the lifetime of the stack frame this Slot’s storage is allocated on.

See slot!() and the module documentation.

Implementations

impl<'frame, T> Slot<'frame, T>

pub unsafe fn new_unchecked( ptr: &'frame mut MaybeUninit<T>, drop_flag: DropFlag<'frame>, ) -> Self

Creates a new Slot with the given pointer as its basis.

To safely construct a Slot, use slot!().

Safety

ptr must not be outlived by any other pointers to its allocation.

drop_flag’s value must be dead, and must be a drop flag governing the destruction of ptr’s storage in an appropriate manner as described in moveit::drop_flag.

pub fn put(self, val: T) -> MoveRef<'frame, T>

Put val into this slot, returning a new MoveRef.

pub fn pin(self, val: T) -> Pin<MoveRef<'frame, T>>

Pin val into this slot, returning a new, pinned MoveRef.

pub fn emplace<N: New<Output = T>>(self, new: N) -> Pin<MoveRef<'frame, T>>

Emplace new into this slot, returning a new, pinned MoveRef.

pub fn try_emplace<N: TryNew<Output = T>>( self, new: N, ) -> Result<Pin<MoveRef<'frame, T>>, N::Error>

Try to emplace new into this slot, returning a new, pinned MoveRef.

pub fn into_pinned(self) -> Slot<'frame, Pin<T>>

Converts this into a slot for a pinned T.

This is safe, since this Slot owns the referenced data, and Pin is explicitly a repr(transparent) type.

pub unsafe fn cast<U>(self) -> Slot<'frame, U>

Converts this Slot from being a slot for a T to being a slot for some other type U.

moveit::slot!(place: u32);
let foo: MoveRef<u16> = unsafe { place.cast::<u16>() }.put(42);

Safety

T must have at least the size and alignment as U.

impl<'frame, T> Slot<'frame, Pin<T>>

pub fn into_unpinned(self) -> Slot<'frame, T>

Converts this into a slot for an unpinned T.

This is safe, since this Slot owns the referenced data, and Pin is explicitly a repr(transparent) type.

Moreover, no actual unpinning is occurring: the referenced data must be uninitialized, so it cannot have a pinned referent.